Normative: Type check calls to methods
This patch rewrites the private method and accessor specification
to make private methods and accessors be non-writable own/static fields.
These fields are added before ordinary fields are added, making private
methods accessible from the initializers of both static and instance
fields.

The patch also organizes the specification in a way that makes it very
friendly to the insertion of decorators. There are a number of "dead"
spec paths which will only be accessible with the addition of decorators.

Closes #1, #3