Normative: Disallow private static methods
At the November 2017 TC39 meeting, the committee decided to back private
static fields up to Stage 2, to be pursued in a separate proposal. This
patch makes use of private static methods a syntax error. However, it
preserves the factoring in this proposal which would make them easy
to add as a follow-on in the future.